Solve jd01_10 Level A, Level B, Level C

+++ src/by/it/chumak/jd01_10/Bean.java
@@ -9,12 +9,12 @@ class Bean {
 
     @Param(a = 5, b = 10)
     double max(int a, int b) {
-        return (double) Math.max(a, b);
+        return Math.max(a, b);
     }
 
     @Param(a = 5, b = 10)
     static double min(int a, int b) {
-        return (double) Math.min(a, b);
+        return Math.min(a, b);
     }
 
     double avg(int a, int b) {
